requirements.txt
python-telegram-bot==21.4
httpx>=0.27
pydantic>=2.7
SQLAlchemy>=2.0
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.30
python-multipart>=0.0.9
